+++ src/Event/FailCallEvent.php
@@ -2,7 +2,8 @@
 
 namespace Biplane\YandexDirect\Event;
 
-use Biplane\YandexDirect\Api\SoapClient;
+use Biplane\YandexDirect\Api\Units;
+use Biplane\YandexDirect\ClientInterface;
 use Biplane\YandexDirect\User;
 
 /**
@@ -17,15 +18,22 @@ class FailCallEvent extends BaseAfterCallEvent
     /**
      * Constructor.
      *
-     * @param string     $methodRef The fullname of API method
-     * @param array      $params    The params for method of API
-     * @param User       $user      The configuration
-     * @param SoapClient $client    The SOAP client
-     * @param \Exception $exception The thrown exception
+     * @param string          $methodRef The fullname of API method
+     * @param array           $params    The params for method of API
+     * @param User            $user      The configuration
+     * @param ClientInterface $client    The client for API service
+     * @param \Exception      $exception The thrown exception
+     * @param Units|null      $units     Information of units
      */
-    public function __construct($methodRef, array $params, User $user, SoapClient $client, \Exception $exception)
-    {
-        parent::__construct($methodRef, $params, $user, $client);
+    public function __construct(
+        $methodRef,
+        array $params,
+        User $user,
+        ClientInterface $client,
+        \Exception $exception,
+        Units $units = null
+    ) {
+        parent::__construct($methodRef, $params, $user, $client, $units);
 
         $this->exception = $exception;
     }