+++ src/Event/PostCallEvent.php
@@ -2,7 +2,8 @@
 
 namespace Biplane\YandexDirect\Event;
 
-use Biplane\YandexDirect\Api\SoapClient;
+use Biplane\YandexDirect\Api\Units;
+use Biplane\YandexDirect\ClientInterface;
 use Biplane\YandexDirect\User;
 
 /**
@@ -17,17 +18,24 @@ class PostCallEvent extends BaseAfterCallEvent
     /**
      * Constructor.
      *
-     * @param string     $methodRef The fullname of API method
-     * @param array      $params    The params for method of API
-     * @param User       $user      The user
-     * @param SoapClient $client    The client
-     * @param mixed      $response  The result from API
+     * @param string          $methodRef The fullname of API method
+     * @param array           $params    The params for method of API
+     * @param User            $user      The user
+     * @param ClientInterface $client    The client for API service
+     * @param mixed           $response  The result from API
+     * @param Units|null      $units     Information of units
      *
      * @throws \InvalidArgumentException
      */
-    public function __construct($methodRef, array $params, User $user, SoapClient $client, $response)
-    {
-        parent::__construct($methodRef, $params, $user, $client);
+    public function __construct(
+        $methodRef,
+        array $params,
+        User $user,
+        ClientInterface $client,
+        $response,
+        Units $units = null
+    ) {
+        parent::__construct($methodRef, $params, $user, $client, $units);
 
         $this->result = $response;
     }